refactor fragment, dialog, recyclerview tests into /ui
